Fix tests for NodeControlTool (#940) (#455)
* Fix tests for NodeControlTool (#940)


* Remove Copyright text (#940)
